mental note: the reason switch to multiple rooms
the reason we switch to multiple rooms is that then we don't have to
worry about locks etc. note how in paper konidaris says that coords are
rel to current room. implies to me that this is how he implemented it as
well (that is to say, with multiple rooms).